+++ src/ledger_editor/utils/__init__.py
@@ -2,6 +2,6 @@
 
 from ledger_editor.utils.date_parser import parse_date
 from ledger_editor.utils.file_resolver import resolve_journal_file
-from ledger_editor.utils.ledger_io import load_journal, save_journal
+from ledger_editor.utils.ledger_io import align_posting_amounts, load_journal, save_journal
 
-__all__ = ["load_journal", "parse_date", "resolve_journal_file", "save_journal"]
+__all__ = ["align_posting_amounts", "load_journal", "parse_date", "resolve_journal_file", "save_journal"]